feat: Split extensions into dedicated entry points
