+++ src/content/examples/custom-protocol/CustomProtocol.svelte
@@ -6,9 +6,9 @@
 		RasterLayer,
 		LineLayer,
 		FillLayer,
-		Protocol,
-		PMTilesProtocol
+		Protocol
 	} from 'svelte-maplibre-gl';
+	import { PMTilesProtocol } from 'svelte-maplibre-gl/pmtiles';
 	import maplibregl from 'maplibre-gl';
 
 	const myProtocolLoadFn: maplibregl.AddProtocolAction = async (params) => {
