+++ src/lib/maplibre/ext/contour/index.ts
@@ -1,0 +1,1 @@
+export { default as MapLibreContourSource } from './MapLibreContourSource.svelte';
