+++ src/lib/maplibre/ext/deckgl/index.ts
@@ -1,0 +1,1 @@
+export { default as DeckGLOverlay } from './DeckGLOverlay.svelte';
