+++ src/lib/maplibre/ext/pmtiles/index.ts
@@ -1,0 +1,1 @@
+export { default as PMTilesProtocol } from './PMTilesProtocol.svelte';
