+++ svelte.config.js
@@ -36,6 +36,9 @@ const config = {
 		adapter: adapter(),
 		alias: {
 			$content: 'src/content',
+			'svelte-maplibre-gl/contour': 'src/lib/maplibre/ext/contour/index.js',
+			'svelte-maplibre-gl/deckgl': 'src/lib/maplibre/ext/deckgl/index.js',
+			'svelte-maplibre-gl/pmtiles': 'src/lib/maplibre/ext/pmtiles/index.js',
 			'svelte-maplibre-gl': 'src/lib/maplibre/index.js'
 		}
 	},
